backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
paho-mqtt>=1.6.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
aiosqlite>=0.19.0
